+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-CustomTemplates-resttemplate/src/main/resources/templates/resttemplate/client_query_mutation_type.vm.java
@@ -71,7 +71,7 @@ import com.graphql_java_generator.client.GraphqlClientUtils;
 @Deprecated
 #end
 ${object.annotation}
-public class ${object.classSimpleName} extends ${object.classSimpleName}Executor #if(!${configuration.separateUtilityClasses} && ${object.requestType})implements com.graphql_java_generator.client.response.GraphQLRequestObject #end{
+public class ${object.classSimpleName} extends ${object.classSimpleName}Executor #if(!${configuration.separateUtilityClasses} && ${object.requestType})implements com.graphql_java_generator.client.GraphQLRequestObject #end{
 
 #if(!${configuration.separateUtilityClasses})
 ##